pyproject.toml
[project]
name = "humanoid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "b2>=4.6.0",
    "lerobot[unitree-g1]>=0.5.1",
    "loguru>=0.7.3",
    "msgpack>=1.1.2",
    "msgpack-numpy>=0.4.8",
    "mujoco>=3.7.0",
    "pinocchio>=0.4.3",
    "robojudo",
    "unitree-rl-mjlab",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.uv.sources]
unitree-rl-mjlab = { git = "https://github.com/unitreerobotics/unitree_rl_mjlab.git" }
robojudo = { git = "https://github.com/HansZ8/RoboJuDo.git" }
